test: update compression test to use full 256 bit range

+++ pkg/pool-weighted/test/foundry/ValueCompression.t.sol
@@ -61,7 +61,7 @@ contract ValueCompressionTest is Test {
     }
 
     function testCompression(
-        uint64 value,
+        uint256 value,
         uint8 bitLength,
         uint256 maxUncompressedValue
     ) external {
